Restructure directories and fix conflicts
