+++ packages/esm-patient-chart-app/src/clinical-views/encounter-list/encounter-list-tabs.component.tsx
