+++ packages/esm-patient-chart-app/src/clinical-views/encounter-list/encounter-list.scss
