+++ packages/esm-patient-chart-app/src/clinical-views/encounter-list/table.component.tsx
