+++ packages/esm-patient-chart-app/src/clinical-views/encounter-list/table.scss
