+++ packages/esm-patient-chart-app/src/clinical-views/encounter-tile/tile.scss
