+++ packages/esm-patient-chart-app/src/clinical-views/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useEncounterRows';
 export * from './useFormsJson';
+export * from './useLastEncounter';
