+++ packages/esm-patient-chart-app/src/clinical-views/hooks/useEncounterRows.ts
